task 3

+++ 0x10-variadic_functions/3-print_all.c
@@ -23,7 +23,7 @@ void print_char(va_list arg)
 
 /**
  * print_int - Prints an int.
- * @arg: A list of argument pointing to the 
+ * @arg: A list of argument pointing to the
  * int to be printed.
  */
 void print_int(va_list arg)
@@ -36,14 +36,14 @@ void print_int(va_list arg)
 
 /**
  * print_float - Prints a float.
- * @arg: A list of argument pointing to the 
+ * @arg: A list of argument pointing to the
  * float to be printed.
  */
 void print_float(va_list arg)
 {
 	float num;
 
-	num = va_arg(arg, int);
+	num = va_arg(arg, double);
 	printf("%f", num);
 }
